+++ README.md
@@ -57,7 +57,7 @@ We got you covered 👉 [stimulus-use.github.io/stimulus-use](https://stimulus-u
   |[`useIntersection`](./docs/use-intersection.md) | Tracks the element's intersection and adds **appear**, **disappear** callbacks to your controller.|`appear`</br> `disappear`|
   |[`useMutation`](./docs/use-mutation.md) | Tracks mutations on an element, its attributes and/or subtree. Adds a **mutate** callback to your controller.|`mutate`|
   |[`useResize`](./docs/use-resize.md)|Tracks the element's size and adds a new lifecycle callback **resize**.|`resize`|
-  |[`useTargetChanges`](./docs/use-target-changes.md) | Tracks when targets are added or removed from the controller's scope. Adds **[target]TargetAdded**  and **[target]TargetRemoved** callback to your controller for each specified target.| `[target]TargetAdded` `[target]TargetRemoved`|
+  |[`useTargetMutation`](docs/use-target-mutation.md) | Tracks when targets are added or removed from the controller's scope. Adds **[target]TargetAdded**  and **[target]TargetRemoved** callback to your controller for each specified target.| `[target]TargetAdded` `[target]TargetRemoved`|
   |[`useVisibility`](./docs/use-visibility.md) </br>| Tracks the page visibility and adds **visible**, **invisible** callbacks to your controller.|`visible`</br> `invisible`|
   |[`useWindowResize`](./docs/use-window-resize.md)| Tracks the size of the `window` object and adds a new lifecycle callback **windowResize**.|`windowResize`|
