+++ src/index.ts
@@ -7,7 +7,7 @@ export { useIdle, IdleController } from './use-idle/index'
 export { useVisibility, UseVisibility, VisibilityController } from './use-visibility/index'
 export { useHover, UseHover, HoverController } from './use-hover/index'
 export { useMutation, UseMutation, MutationController } from './use-mutation/index'
-export { useTargetChanges, UseTargetChanges, TargetChangesController } from './use-target-changes/index'
+export { useTargetMutation, UseTargetMutation, TargetMutationController } from './use-target-mutation/index'
 export { useWindowResize, WindowResizeController } from './use-window-resize/index'
 export { useMemo } from './use-memo/index'
 export { useDebounce } from './use-debounce/index'
