--- src/use-target-changes/index.ts
@@ -1,2 +1,0 @@
-export { useTargetChanges, UseTargetChanges } from './use-target-changes'
-export { TargetChangesController } from './target-changes-controller'
