+++ src/use-target-mutation/index.ts
@@ -1,0 +1,2 @@
+export { useTargetMutation, UseTargetMutation } from './use-target-mutation'
+export { TargetMutationController } from './target-mutation-controller'
